Move some utils into ReactFXUtil
